changed text and commands to refer to NWM
Future edits: https://hub.docker.com/r/wrfhydro/training/ this link will need to change to the actual link to the private container. 
Someone will actually need to make sure that the commands work or change them 
We might need to specify the branch of the training repo to post issues to. I put this in https://github.com/NCAR/wrf_hydro_training/edit/RFC_Training/
The password for the web browser is currently wrfhydrotraining this will need to change to nwmtraining